- removed BcscAuthExtension.kt - added AppAuthExtensions.kt to appAuth module
