+++ appauth/build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id 'com.android.library'
+    id 'kotlin-android'
 }
 
 android {
@@ -20,17 +21,45 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
+    flavorDimensions "version"
+    productFlavors {
+
+        mock {
+            dimension "version"
+        }
+        dev {
+            dimension "version"
+        }
+
+        stage {
+            dimension "version"
+        }
+
+        prod {
+            dimension "version"
+        }
+    }
+
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
     }
 }
 
 dependencies {
 
     implementation "androidx.appcompat:appcompat:$versions.appcompat"
+
+    implementation project(path: ':common')
+
     api "androidx.browser:browser:$versions.browser"
     implementation "androidx.annotation:annotation:$versions.browser"
+
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
