+++ common/src/main/java/ca/bc/gov/common/exceptions/MyHealthAuthException.kt
@@ -1,0 +1,8 @@
+package ca.bc.gov.common.exceptions
+
+import java.io.IOException
+
+/**
+ * @author Pinakin Kansara
+ */
+class MyHealthAuthException(message: String?) : IOException(message)
