+++ common/src/main/java/ca/bc/gov/common/model/BCSCAuthDataDto.kt
@@ -1,0 +1,9 @@
+package ca.bc.gov.common.model
+
+/**
+ * @author Pinakin Kansara
+ */
+data class BCSCAuthDataDto(
+    val authToken: String,
+    val hdId: String
+)
